+++ src/ij/io/BitBuffer.java
@@ -36,7 +36,7 @@ public class BitBuffer {
 				if (currentBit == 0) { // special
 					toStore = toStore << 8;
 					int cb = ((int) byteBuffer[currentByte]);
-					toStore += (cb<0 ? (int) 256 + cb : (int) cb);
+					toStore += (cb<0 ? 256 + cb : cb);
 					bitsToRead -= 8;
 					currentByte++;
 				} else {
@@ -49,7 +49,7 @@ public class BitBuffer {
 			} else {
 				toStore = toStore << bitsToRead;
 				int cb = ((int) byteBuffer[currentByte]);
-				cb = (cb<0 ? (int) 256 + cb : (int) cb);
+				cb = (cb<0 ? 256 + cb : cb);
 				toStore += ((cb) & (0x00FF - frontMask[currentBit])) >> (8 - (currentBit + bitsToRead));
 				currentBit += bitsToRead;
 				bitsToRead = 0;
